Checkpointing more work
Things run pretty well now. The linearized parsing seems to be fairly
easy to work with so far for at least the simple checks. Off for lunch
and then will try and put together a more difficult check for proper
indentation checking. Assuming that's as easy as I think it will be then
I'll spend some effort on the front end to make the output prettier as
well as support some basic command line options to tweak how things run.

I also need to figure out some method for ignoring specific files and/or
lines and/or errors.