Handle errors while parsing source files
This handles when a source file can not be parsed correctly. This can
either be due to actual syntax errors or because of some specific macro
arguments that are not complete expressions (i.e., ?assertMatch with a
guard). For the former, we'll always have to deal with the situation.
For macros we may want to eventually try and do some sort of
intermediate and/or skip parsing retry if we find an error.