Replace unparseable define and macro arguments
Arguments to macros and defines don't actually have to be syntactically
valid as they get pre-processed away normally. Rather than run a whole
preprocessor step (which would be counter productive for us) this just
attempts to parse argument expressions for -define attributes and macro
calls. Any argument that can't be parsed is replaced with a stringified
version of itself.